import++

+++ src/org/footware/server/db/Track.java
@@ -65,7 +65,7 @@ public class Track implements Serializable {
 	
 	private boolean commentsEnabled;
 	private int trackpoints;
-	private int length;
+	private double length;
 	private double midLongitude;
 	private double midLatitude;
 	private Date startTime;
@@ -222,7 +222,7 @@ public class Track implements Serializable {
 	 * Gets the length in meters (TODO: Check unit) of this track
 	 * @return the length in meters of this track
 	 */
-	public int getLength() {
+	public double getLength() {
 		return length;
 	}
 	
@@ -231,7 +231,7 @@ public class Track implements Serializable {
 	 * Should probably not be used from the client side
 	 * @param length new length in meters for this track
 	 */
-	public void setLength(int length) {
+	public void setLength(double length) {
 		this.length = length;
 	}
